Added two other things to Alt + X
Added in hiding of Indicator and Balloons. Not what I initially thought they were but they're called with Ctrl + X and don't seem to have any hinderance on experience.

With those additions the search part for SetEnableNamePlate was changed to resolve the full address of the function from its relative call.